Allow calculations to use floats

+++ datagrid/core.py
@@ -102,6 +102,6 @@ class DataGrid(object):
         if len(self.aggregatemethods):
             columnValues = zip(*data)
             for i, m in self.aggregatemethods.items():
-                rowData[i] = str(m([int(v) for v in columnValues[i] if v != '']))
+                rowData[i] = str(m([float(v) for v in columnValues[i] if v != '']))
         return rowData
